Fix Delta 3 missing DC 12V port
